Refactor : 검색어 관리 방식 useSearchParams → useState로 변경
